Fixed crashing of map data and search result table
